+++ lib/models/indicator_summary_model.dart
@@ -1,7 +1,7 @@
 class IndicatorSummaryModel {
   final String label;
   final String value;
-  final int? score;
+  final double? score;
   final String? status;
 
   const IndicatorSummaryModel({required this.label, required this.value, this.score, this.status});
