+++ lib/models/weighted_score_model.dart
@@ -1,5 +1,5 @@
 class WeightedScore {
-  final int? score;
+  final double? score;
   final double weight;
 
   const WeightedScore({required this.score, required this.weight});
